refactor: 이미지 업로더 컴포넌트 분리
